Refactor, Fix bugs
  - Change int on long
  - Fix mapping of ORM
  - Change internal modified to public in some place

+++ SearchEngine/Application/Companies/ICompanyService.cs
@@ -4,6 +4,6 @@ namespace Aveneo.SearchEngine.Application.Companies
 {
     internal interface ICompanyService
     {
-        Maybe<CompanyResult> GetCompanyByPredicate(int id);
+        Maybe<CompanyResult> GetCompanyByPredicate(long id);
     }
 }
